requirements/prod.txt
-r base.txt

django-debug-toolbar
ipython
black
flake8